Clean up entity logic
